Rework wad info to dynamic object
This is for future plans to allow compilation to verify that everything has been assembled correctly. The wad project will be stored in a global object, both to make it easier to format AND easier to serialize to JSON. Not ideal for fast programs, but this is an IDE of sorts, so it's speed is not an as big concern.

Todo: Directories need additional info on them to identify that they are directories when read. Maybe check for a lump array object?